refactor callback thread handling
